+++ tests/Unit/Executor/ParameterSetExtractorTest.php
@@ -10,11 +10,11 @@
  *
  */
 
-namespace PhpBench\Tests\Unit\Benchmark\Executor;
+namespace PhpBench\Tests\Unit\Executor;
 
 use PhpBench\Benchmark\Remote\Exception\ScriptErrorException;
 use PhpBench\Benchmark\Remote\Payload;
-use PhpBench\Tests\Unit\Benchmark\Executor\benchmarks\ParamProviderBench;
+use PhpBench\Tests\Unit\Executor\benchmarks\ParamProviderBench;
 use PHPUnit\Framework\TestCase;
 
 class ParameterSetExtractorTest extends TestCase